requirements.txt
Django==2.0.1
psycopg2==2.7.3.2
djangorestframework==3.7.7
django-pagedown==1.0.4
django-markup[all-filter-dependencies]
markdown
django-cors-headers==2.1.0
